requirements.txt
pyyaml
psutil
colorama
pyinquirer
